+++ src/lib/Components/Bidding/Screens/BillingAddress.tsx
@@ -2,7 +2,16 @@ import React from "react"
 
 import { Schema, screenTrack, track } from "../../../utils/track"
 
-import { Dimensions, KeyboardAvoidingView, NavigatorIOS, ScrollView, TouchableWithoutFeedback } from "react-native"
+import {
+  Dimensions,
+  EmitterSubscription,
+  Keyboard,
+  KeyboardAvoidingView,
+  LayoutRectangle,
+  NavigatorIOS,
+  ScrollView,
+  TouchableWithoutFeedback,
+} from "react-native"
 
 import { Flex } from "../Elements/Flex"
 import { Sans12, Serif16 } from "../Elements/Typography"
@@ -27,13 +36,15 @@ interface StyledInputInterface {
   }
 }
 
-const StyledInput = ({ label, error, ...props }) => (
-  <Flex mb={4}>
+const StyledInput = ({ label, error, onLayout, ...props }) => (
+  <Flex mb={4} onLayout={onLayout}>
     <Serif16 mb={2}>{label}</Serif16>
     <Input mb={3} error={Boolean(error)} {...props} />
     {error && <Sans12 color="red100">{error}</Sans12>}
   </Flex>
 )
+
+const iOSAccessoryViewHeight = 60
 
 interface BillingAddressProps {
   onSubmit?: (values: Address) => void
@@ -64,6 +75,19 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
   private city: StyledInputInterface
   private stateProvinceRegion: StyledInputInterface
   private postalCode: StyledInputInterface
+
+  private fullNameLayout: LayoutRectangle
+  private addressLine1Layout: LayoutRectangle
+  private addressLine2Layout: LayoutRectangle
+  private cityLayout: LayoutRectangle
+  private stateProvinceRegionLayout: LayoutRectangle
+  private postalCodeLayout: LayoutRectangle
+
+  private keyboardDidShowListener: EmitterSubscription
+
+  private keyboardHeight: number
+
+  private scrollView: ScrollView
 
   constructor(props) {
     super(props)
@@ -135,19 +159,26 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
     })
   }
 
+  componentWillMount() {
+    this.keyboardDidShowListener = Keyboard.addListener(
+      "keyboardDidShow",
+      ({ endCoordinates }) => (this.keyboardHeight = endCoordinates.height)
+    )
+  }
+
+  componentWillUnmount() {
+    this.keyboardDidShowListener.remove()
+  }
+
   render() {
     const errorForCountry = this.state.errors.country
 
-    // TODO: Remove this once React Native has been updated
-    const isPhoneX = Dimensions.get("window").height === 812 && Dimensions.get("window").width === 375
-    const defaultVerticalOffset = isPhoneX ? 30 : 15
-
     return (
       <BiddingThemeProvider>
-        <KeyboardAvoidingView behavior="padding" keyboardVerticalOffset={defaultVerticalOffset} style={{ flex: 1 }}>
+        <KeyboardAvoidingView behavior="padding" keyboardVerticalOffset={this.verticalOffset} style={{ flex: 1 }}>
           <BackButton navigator={this.props.navigator} />
 
-          <ScrollView>
+          <ScrollView ref={scrollView => (this.scrollView = scrollView as any)}>
             <Container>
               <Title mt={0} mb={6}>
                 Your billing address
@@ -159,6 +190,8 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
                 placeholder="Add your full name"
                 autoFocus={true}
                 onSubmitEditing={() => this.addressLine1.root.focus()}
+                onLayout={({ nativeEvent }) => (this.fullNameLayout = nativeEvent.layout)}
+                onFocus={() => this.scrollView.scrollTo({ x: 0, y: this.yPosition(this.fullNameLayout) })}
               />
 
               <StyledInput
@@ -166,6 +199,8 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
                 label="Address line 1"
                 placeholder="Add your street address"
                 onSubmitEditing={() => this.addressLine2.root.focus()}
+                onLayout={({ nativeEvent }) => (this.addressLine1Layout = nativeEvent.layout)}
+                onFocus={() => this.scrollView.scrollTo({ x: 0, y: this.yPosition(this.addressLine1Layout) })}
               />
 
               <StyledInput
@@ -173,6 +208,8 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
                 label="Address line 2 (optional)"
                 placeholder="Add your apt, floor, suite, etc."
                 onSubmitEditing={() => this.city.root.focus()}
+                onLayout={({ nativeEvent }) => (this.addressLine2Layout = nativeEvent.layout)}
+                onFocus={() => this.scrollView.scrollTo({ x: 0, y: this.yPosition(this.addressLine2Layout) })}
               />
 
               <StyledInput
@@ -180,6 +217,8 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
                 label="City"
                 placeholder="Add your city"
                 onSubmitEditing={() => this.stateProvinceRegion.root.focus()}
+                onLayout={({ nativeEvent }) => (this.cityLayout = nativeEvent.layout)}
+                onFocus={() => this.scrollView.scrollTo({ x: 0, y: this.yPosition(this.cityLayout) })}
               />
 
               <StyledInput
@@ -188,6 +227,13 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
                 placeholder="Add state, province, or region"
                 onSubmitEditing={() => this.postalCode.root.focus()}
                 inputRef={el => (this.stateProvinceRegion = el)}
+                onLayout={({ nativeEvent }) => (this.stateProvinceRegionLayout = nativeEvent.layout)}
+                onFocus={() =>
+                  this.scrollView.scrollTo({
+                    x: 0,
+                    y: this.yPosition(this.stateProvinceRegionLayout),
+                  })
+                }
               />
 
               <StyledInput
@@ -195,6 +241,8 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
                 label="Postal code"
                 placeholder="Add your postal code"
                 onSubmitEditing={() => this.presentSelectCountry()}
+                onLayout={({ nativeEvent }) => (this.postalCodeLayout = nativeEvent.layout)}
+                onFocus={() => this.scrollView.scrollTo({ x: 0, y: this.yPosition(this.postalCodeLayout) })}
               />
 
               <Flex mb={4}>
@@ -231,5 +279,22 @@ export class BillingAddress extends React.Component<BillingAddressProps, Billing
       returnKeyType: "next",
       value: this.state.values[field],
     }
+  }
+
+  private yPosition({ y, height }) {
+    const windowHeight = Dimensions.get("window").height
+
+    return Math.max(0, y - windowHeight + height + iOSAccessoryViewHeight + this.keyboardHeight + this.iPhoneXOffset)
+  }
+
+  private get verticalOffset() {
+    return this.iPhoneXOffset + 15
+  }
+
+  // TODO: Remove this once React Native has been updated
+  private get iPhoneXOffset() {
+    const isPhoneX = Dimensions.get("window").height === 812 && Dimensions.get("window").width === 375
+
+    return isPhoneX ? 15 : 0
   }
 }